Store the tiles on the tile server.
I'd forgotten that when I was working on this last year, I'd made the
decision to store the rendered tiles inside the web service container.
In hindsight that doesn't really make any sense, and it doesn't really
work very well.

This change allows the tile server to use its own cache of rendered
tiles stored in the osm-tiles volume.