Clicking clusters launches modals with accident info.
Extensive updates to the heatmap to support launching Bootstrap modals
when clusters containing individual items are clicked.

Also created a backend /get_accident route which retrieves the requested
accident information from the accident dictionary and returns it to the
template.

Lastly modified the process_acc_dict function to return a coords_dict
instead of a coords_list so that accident IDs can be included with the
coordinates.