Implement a redis service for caching the acc_dict
Due to poor design choices I made when starting this project, the
acc_dict is contained in a large JSON file, growing by ~5.5 MB per year.

While it might be best to convert the JSON into an actual database, it
seems simpler to create a redis service for caching and serving it.