Configuring vm.overcommit-memory = 1 in the host environment
I hadn't understood that the vm.overcommit_memory changes need to be made
on the host system rather than in the container. As such, there's no need
for a convoluted runtime command.

Also significantly uping the mem_limit for the container. If this doesn't
do it I might need a beefier VPS.